rename submission file from songsubmission to votesubmission. Updated imports of component
