+++ client/src/components/submissions/VoteSubmission.jsx
